Major CMake refactor
