+++ cmake/install-config.cmake.in
@@ -7,3 +7,31 @@ set_and_check(sleigh_INSTALL_SPECDIR "@PACKAGE_sleigh_INSTALL_SPECDIR@")
 
 # Helpers exposed by default when finding sleigh
 include("${CMAKE_CURRENT_LIST_DIR}/modules/sleighCompile.cmake")
+
+# Find-package component handling
+set(_sleigh_supported_components Support ExtraTools)
+
+set(_sleigh_Support_targets "${CMAKE_CURRENT_LIST_DIR}/sleighSupportTargets.cmake")
+set(_sleigh_ExtraTools_targets "${CMAKE_CURRENT_LIST_DIR}/sleighExtraToolsTargets.cmake")
+
+macro(sleigh_load_component comp)
+  message(STATUS "Trying to load component: ${comp}")
+  if(NOT EXISTS "${_sleigh_${comp}_targets}" AND ${sleigh_FIND_REQUIRED_${comp}})
+    set(sleigh_NOT_FOUND_MESSAGE
+      "sleigh `${comp}` component targets were requested but not found.")
+    set(sleigh_FOUND FALSE)
+    return()
+  endif()
+  if(EXISTS "${_sleigh_${comp}_targets}")
+    include("${_sleigh_${comp}_targets}")
+  endif()
+endmacro()
+
+foreach(_comp ${sleigh_FIND_COMPONENTS})
+  if (NOT _comp IN_LIST _sleigh_supported_components)
+    set(sleigh_FOUND False)
+    set(sleigh_NOT_FOUND_MESSAGE "Unsupported component: ${_comp}")
+  endif()
+  message(STATUS "Searching for component: ${_comp}")
+  sleigh_load_component("${_comp}")
+endforeach()
