+++ cmake/options.cmake
@@ -6,26 +6,56 @@
 # the LICENSE file found in the root directory of this source tree.
 #
 
-option(sleigh_ENABLE_TESTS "Set to true to enable tests" ON)
-option(sleigh_ENABLE_EXAMPLES "Set to true to build examples" ON)
-option(sleigh_ENABLE_DOCUMENTATION "Set to true to enable the documentation")
-option(sleigh_ENABLE_PACKAGING "Set to true to enable packaging")
-option(sleigh_ENABLE_SANITIZERS "Set to true to enable sanitizers")
+# ---- Developer mode ----
 
-# Internal debug settings
-option(sleigh_OPACTION_DEBUG "Turns on all the action tracing facilities")
-option(sleigh_MERGEMULTI_DEBUG "Check for MULTIEQUAL and INDIRECT intersections")
-option(sleigh_BLOCKCONSISTENT_DEBUG "Check that block graph structure is consistent")
-option(sleigh_DFSVERIFY_DEBUG "Make sure that the block ordering algorithm produces a true depth first traversal of the dominator tree")
-
-# Additional internal settings
-option(sleigh_CPUI_STATISTICS "Turn on collection of cover and cast statistics")
-option(sleigh_CPUI_RULECOMPILE "Allow user defined dynamic rules")
-
-# Sanity checking
-if(CMAKE_HOST_SYSTEM_NAME STREQUAL "Windows")
-  set(sleigh_ENABLE_DOCUMENTATION OFF CACHE BOOL "Unsupported on Windows" FORCE)
+# Developer mode enables targets and code paths in the CMake scripts that are
+# only relevant for the developer(s) of sleigh
+# Targets necessary to build the project must be provided unconditionally, so
+# consumers can trivially build and package the project
+if(PROJECT_IS_TOP_LEVEL)
+  option(sleigh_DEVELOPER_MODE "Enable developer mode" OFF)
+  option(BUILD_SHARED_LIBS "Build shared libs. (Untested and not supported)" OFF)
 endif()
+
+# Add-ons by ToB
+option(sleigh_BUILD_SUPPORT "Build ToB support libraries")
+option(sleigh_BUILD_EXTRATOOLS "Build extra ToB sleigh tools")
+if(sleigh_BUILD_EXTRATOOLS)
+  set(sleigh_BUILD_SUPPORT ON CACHE BOOL "Build ToB support libraries" FORCE)
+endif()
+
+# Internal settings
+option(sleigh_CPUI_RULECOMPILE "Allow user defined dynamic rules")
+option(sleigh_CPUI_STATISTICS "Turn on collection of cover and cast statistics")
+
+# Internal debug settings (naming is swapped to help with discoverability in CMake options)
+option(sleigh_DEBUG_BLOCKCONSISTENT "Check that block graph structure is consistent")
+option(sleigh_DEBUG_DFSVERIFY "Make sure that the block ordering algorithm produces a true depth first traversal of the dominator tree")
+option(sleigh_DEBUG_MERGEMULTI "Check for MULTIEQUAL and INDIRECT intersections")
+option(sleigh_DEBUG_OPACTION "Turns on all the action tracing facilities")
+
+macro(sleigh_add_optional_defines target visibility)
+  set(opt_defines "")
+  if(sleigh_CPUI_RULECOMPILE)
+    list(APPEND opt_defines "CPUI_RULECOMPILE")
+  endif()
+  if(sleigh_CPUI_STATISTICS)
+    list(APPEND opt_defines "CPUI_STATISTICS")
+  endif()
+  if(sleigh_DEBUG_BLOCKCONSISTENT)
+    list(APPEND opt_defines "BLOCKCONSISTENT_DEBUG")
+  endif()
+  if(sleigh_DEBUG_DFSVERIFY)
+    list(APPEND opt_defines "DFSVERIFY_DEBUG")
+  endif()
+  if(sleigh_DEBUG_MERGEMULTI)
+    list(APPEND opt_defines "MERGEMULTI_DEBUG")
+  endif()
+  if(sleigh_DEBUG_OPACTION)
+    list(APPEND opt_defines "OPACTION_DEBUG")
+  endif()
+  target_compile_definitions("${target}" ${visibility} ${opt_defines})
+endmacro()
 
 # ---- Warning guard ----
