--- include/libsleigh.hh
@@ -1,43 +1,0 @@
-/*
-  Copyright (c) 2021-present, Trail of Bits, Inc.
-  All rights reserved.
-
-  This source code is licensed in accordance with the terms specified in
-  the LICENSE file found in the root directory of this source tree.
-*/
-
-#pragma once
-#ifndef _MSC_VER
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wdeprecated"
-#pragma GCC diagnostic ignored "-Wsign-compare"
-#pragma GCC diagnostic ignored "-Wunused-parameter"
-#endif
-#include "address.hh"
-#include "context.hh"
-#include "emulate.hh"
-#include "error.hh"
-#include "float.hh"
-#include "globalcontext.hh"
-#include "loadimage.hh"
-#include "memstate.hh"
-#include "opbehavior.hh"
-#include "opcodes.hh"
-#include "partmap.hh"
-#include "pcoderaw.hh"
-#include "semantics.hh"
-#include "sleigh.hh"
-#include "sleighbase.hh"
-#include "slghpatexpress.hh"
-#include "slghpattern.hh"
-#include "slghsymbol.hh"
-#include "space.hh"
-#include "translate.hh"
-#include "types.h"
-#include "xml.hh"
-#ifndef _MSC_VER
-#pragma GCC diagnostic pop
-#endif
-
-#include "Support.h"
-#include "Version.h"
