+++ support/include/sleigh/libsleigh.hh
@@ -1,0 +1,43 @@
+/*
+  Copyright (c) 2021-present, Trail of Bits, Inc.
+  All rights reserved.
+
+  This source code is licensed in accordance with the terms specified in
+  the LICENSE file found in the root directory of this source tree.
+*/
+
+#pragma once
+#ifndef _MSC_VER
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wdeprecated"
+#pragma GCC diagnostic ignored "-Wsign-compare"
+#pragma GCC diagnostic ignored "-Wunused-parameter"
+#endif
+#include "sleigh/address.hh"
+#include "sleigh/context.hh"
+#include "sleigh/emulate.hh"
+#include "sleigh/error.hh"
+#include "sleigh/float.hh"
+#include "sleigh/globalcontext.hh"
+#include "sleigh/loadimage.hh"
+#include "sleigh/memstate.hh"
+#include "sleigh/opbehavior.hh"
+#include "sleigh/opcodes.hh"
+#include "sleigh/partmap.hh"
+#include "sleigh/pcoderaw.hh"
+#include "sleigh/semantics.hh"
+#include "sleigh/sleigh.hh"
+#include "sleigh/sleighbase.hh"
+#include "sleigh/slghpatexpress.hh"
+#include "sleigh/slghpattern.hh"
+#include "sleigh/slghsymbol.hh"
+#include "sleigh/space.hh"
+#include "sleigh/translate.hh"
+#include "sleigh/types.h"
+#include "sleigh/xml.hh"
+#ifndef _MSC_VER
+#pragma GCC diagnostic pop
+#endif
+
+#include "sleigh/Support.h"
+#include "sleigh/Version.h"
